+++ frontend/src/components/Sidebar/HeadlampButton.tsx
@@ -1,13 +1,9 @@
 import Button from '@material-ui/core/Button';
 import { makeStyles } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
-import { isValidElement } from 'react';
 import { useTranslation } from 'react-i18next';
 import { getThemeName } from '../../lib/themes';
-import { useTypedSelector } from '../../redux/reducers/reducers';
-import { EmptyContent } from '../common';
-import ErrorBoundary from '../common/ErrorBoundary';
-import AppLogo from './AppLogo';
+import { AppLogo } from '../App/AppLogo';
 
 const useStyle = makeStyles(theme => ({
   toolbar: {
@@ -55,8 +51,6 @@ export default function HeadlampButton({
   const isSmall = useMediaQuery('(max-width:600px)');
   const classes = useStyle({ isSidebarOpen: open, isSmall: isSmall });
   const { t } = useTranslation('sidebar');
-  const arePluginsLoaded = useTypedSelector(state => state.plugins.loaded);
-  const PluginAppLogoComponent = useTypedSelector(state => state.ui.branding.logo);
 
   if (mobileOnly && (!isSmall || (isSmall && open))) {
     return null;
@@ -70,28 +64,11 @@ export default function HeadlampButton({
         aria-label={open ? t('Shrink sidebar') : t('Expand sidebar')}
         disabled={disabled}
       >
-        <ErrorBoundary>
-          {
-            // Till all plugins are not loaded show empty content for logo as we might have logo coming from a plugin
-            !arePluginsLoaded ? (
-              <EmptyContent />
-            ) : PluginAppLogoComponent ? (
-              isValidElement(PluginAppLogoComponent) ? (
-                // If it's an element, just use it.
-                PluginAppLogoComponent
-              ) : (
-                // It is a component, so we make it here.
-                <PluginAppLogoComponent
-                  logoType={open ? 'large' : 'small'}
-                  themeName={getThemeName()}
-                  className={classes.logo}
-                />
-              )
-            ) : (
-              <AppLogo logoType={open ? 'large' : 'small'} className={classes.logo} />
-            )
-          }
-        </ErrorBoundary>
+        <AppLogo
+          logoType={open ? 'large' : 'small'}
+          themeName={getThemeName()}
+          className={classes.logo}
+        />
       </Button>
     </div>
   );
