+++ frontend/src/components/cluster/Chooser.tsx
@@ -25,16 +25,14 @@ import helpers from '../../helpers';
 import { useClustersConf } from '../../lib/k8s';
 import { Cluster } from '../../lib/k8s/cluster';
 import { createRouteURL } from '../../lib/router';
-import { getThemeName } from '../../lib/themes';
 import { getCluster, getClusterPrefixedPath } from '../../lib/util';
 import { setVersionDialogOpen } from '../../redux/actions/actions';
 import { useTypedSelector } from '../../redux/reducers/reducers';
-import { EmptyContent } from '../common';
+import { AppLogo } from '../App/AppLogo';
 import ActionButton from '../common/ActionButton';
 import { DialogTitle } from '../common/Dialog';
 import ErrorBoundary from '../common/ErrorBoundary';
 import Loader from '../common/Loader';
-import AppLogo from '../Sidebar/AppLogo';
 import ClusterChooser from './ClusterChooser';
 
 export interface ClusterTitleProps {
@@ -278,8 +276,6 @@ export function ClusterDialog(props: ClusterDialogProps) {
   // Only used if open is not provided
   const [show, setShow] = React.useState(true);
   const dispatch = useDispatch();
-  const arePluginsLoaded = useTypedSelector(state => state?.plugins?.loaded);
-  const PluginAppLogoComponent = useTypedSelector(state => state?.ui?.branding?.logo);
 
   function handleClose() {
     if (onClose !== null) {
@@ -322,28 +318,7 @@ export function ClusterDialog(props: ClusterDialogProps) {
           ),
         ]}
       >
-        <ErrorBoundary>
-          {
-            // Till all plugins are not loaded show empty content for logo as we might have logo coming from a plugin
-            !arePluginsLoaded ? (
-              <EmptyContent />
-            ) : PluginAppLogoComponent ? (
-              isValidElement(PluginAppLogoComponent) ? (
-                // If it's an element, just use it.
-                PluginAppLogoComponent
-              ) : (
-                // It is a component, so we make it here.
-                <PluginAppLogoComponent
-                  logoType={'large'}
-                  themeName={getThemeName()}
-                  className={classes.logo}
-                />
-              )
-            ) : (
-              <AppLogo logoType={'large'} className={classes.logo} />
-            )
-          }
-        </ErrorBoundary>
+        <AppLogo logoType={'large'} className={classes.logo} />
       </DialogTitle>
       <DialogContent className={classes.chooserDialog}>{children}</DialogContent>
     </Dialog>
